+++ src/main/kotlin/com/featurevisor/sdk/Emitter.kt
@@ -1,5 +1,7 @@
 package com.featurevisor.sdk
 
+import com.featurevisor.sdk.types.EventName
+
 class Emitter {
 
     private val listeners = mutableMapOf<EventName, (Array<out Any>) -> Unit>()
