+++ src/main/kotlin/com/featurevisor/sdk/Instance+Activation.kt
@@ -1,6 +1,10 @@
 package com.featurevisor.sdk
 
-import com.featurevisor.sdk.EventName.ACTIVATION
+import com.featurevisor.sdk.types.AttributeValue
+import com.featurevisor.sdk.types.Context
+import com.featurevisor.sdk.types.EventName.ACTIVATION
+import com.featurevisor.sdk.types.FeatureKey
+import com.featurevisor.sdk.types.VariationValue
 
 fun FeaturevisorInstance.activate(featureKey: FeatureKey, context: Context = emptyMap()): VariationValue? {
     val evaluation = evaluateVariation(featureKey, context)
