+++ src/main/kotlin/com/featurevisor/sdk/Instance+Evaluation.kt
@@ -2,6 +2,23 @@ package com.featurevisor.sdk
 
 import com.featurevisor.sdk.Conditions.allConditionsAreMatched
 import com.featurevisor.sdk.EvaluationReason.*
+import com.featurevisor.sdk.types.AttributeKey
+import com.featurevisor.sdk.types.AttributeValue
+import com.featurevisor.sdk.types.BucketBy
+import com.featurevisor.sdk.types.BucketKey
+import com.featurevisor.sdk.types.BucketValue
+import com.featurevisor.sdk.types.Context
+import com.featurevisor.sdk.types.Feature
+import com.featurevisor.sdk.types.FeatureKey
+import com.featurevisor.sdk.types.OverrideFeature
+import com.featurevisor.sdk.types.Required
+import com.featurevisor.sdk.types.RuleKey
+import com.featurevisor.sdk.types.Traffic
+import com.featurevisor.sdk.types.VariableKey
+import com.featurevisor.sdk.types.VariableSchema
+import com.featurevisor.sdk.types.VariableValue
+import com.featurevisor.sdk.types.Variation
+import com.featurevisor.sdk.types.VariationValue
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromJsonElement
