+++ src/main/kotlin/com/featurevisor/sdk/Instance+Refresh.kt
@@ -1,6 +1,7 @@
 package com.featurevisor.sdk
 
 import com.featurevisor.sdk.FeaturevisorError.*
+import com.featurevisor.sdk.types.EventName
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
