+++ src/main/kotlin/com/featurevisor/sdk/Instance+Variation.kt
@@ -1,5 +1,9 @@
 package com.featurevisor.sdk
 
+import com.featurevisor.sdk.types.Context
+import com.featurevisor.sdk.types.FeatureKey
+import com.featurevisor.sdk.types.VariationValue
+
 internal fun FeaturevisorInstance.getVariation(featureKey: FeatureKey, context: Context): VariationValue? {
     val evaluation = evaluateVariation(featureKey, context)
     return when {
