+++ src/main/kotlin/com/featurevisor/sdk/types/Types.kt
@@ -1,4 +1,4 @@
-package com.featurevisor.sdk
+package com.featurevisor.sdk.types
 
 import com.featurevisor.sdk.serializers.BucketBySerializer
 import com.featurevisor.sdk.serializers.ConditionSerializer
@@ -351,7 +351,7 @@ data class RequiredWithVariation(
 
 @Serializable
 sealed class Required {
-    data class FeatureKey(val required: com.featurevisor.sdk.FeatureKey) : Required()
+    data class FeatureKey(val required: com.featurevisor.sdk.types.FeatureKey) : Required()
     data class WithVariation(val required: RequiredWithVariation) : Required()
 }
