+++ src/test/kotlin/com/featurevisor/sdk/ConditionsTest.kt
@@ -1,24 +1,27 @@
 package com.featurevisor.sdk
 
-import com.featurevisor.sdk.Operator.AFTER
-import com.featurevisor.sdk.Operator.BEFORE
-import com.featurevisor.sdk.Operator.CONTAINS
-import com.featurevisor.sdk.Operator.ENDS_WITH
-import com.featurevisor.sdk.Operator.EQUALS
-import com.featurevisor.sdk.Operator.GREATER_THAN
-import com.featurevisor.sdk.Operator.GREATER_THAN_OR_EQUAL
-import com.featurevisor.sdk.Operator.IN_ARRAY
-import com.featurevisor.sdk.Operator.LESS_THAN
-import com.featurevisor.sdk.Operator.LESS_THAN_OR_EQUAL
-import com.featurevisor.sdk.Operator.NOT_EQUALS
-import com.featurevisor.sdk.Operator.NOT_IN_ARRAY
-import com.featurevisor.sdk.Operator.SEMVER_EQUALS
-import com.featurevisor.sdk.Operator.SEMVER_GREATER_THAN
-import com.featurevisor.sdk.Operator.SEMVER_GREATER_THAN_OR_EQUAL
-import com.featurevisor.sdk.Operator.SEMVER_LESS_THAN
-import com.featurevisor.sdk.Operator.SEMVER_LESS_THAN_OR_EQUAL
-import com.featurevisor.sdk.Operator.SEMVER_NOT_EQUALS
-import com.featurevisor.sdk.Operator.STARTS_WITH
+import com.featurevisor.sdk.types.AttributeValue
+import com.featurevisor.sdk.types.Condition
+import com.featurevisor.sdk.types.ConditionValue
+import com.featurevisor.sdk.types.Operator.AFTER
+import com.featurevisor.sdk.types.Operator.BEFORE
+import com.featurevisor.sdk.types.Operator.CONTAINS
+import com.featurevisor.sdk.types.Operator.ENDS_WITH
+import com.featurevisor.sdk.types.Operator.EQUALS
+import com.featurevisor.sdk.types.Operator.GREATER_THAN
+import com.featurevisor.sdk.types.Operator.GREATER_THAN_OR_EQUAL
+import com.featurevisor.sdk.types.Operator.IN_ARRAY
+import com.featurevisor.sdk.types.Operator.LESS_THAN
+import com.featurevisor.sdk.types.Operator.LESS_THAN_OR_EQUAL
+import com.featurevisor.sdk.types.Operator.NOT_EQUALS
+import com.featurevisor.sdk.types.Operator.NOT_IN_ARRAY
+import com.featurevisor.sdk.types.Operator.SEMVER_EQUALS
+import com.featurevisor.sdk.types.Operator.SEMVER_GREATER_THAN
+import com.featurevisor.sdk.types.Operator.SEMVER_GREATER_THAN_OR_EQUAL
+import com.featurevisor.sdk.types.Operator.SEMVER_LESS_THAN
+import com.featurevisor.sdk.types.Operator.SEMVER_LESS_THAN_OR_EQUAL
+import com.featurevisor.sdk.types.Operator.SEMVER_NOT_EQUALS
+import com.featurevisor.sdk.types.Operator.STARTS_WITH
 import io.kotest.matchers.shouldBe
 import java.time.LocalDate
 import kotlin.test.Test
